chore: move env variables
